Remove source transformers from map
Now, the transformers for the source cell types are passed in via
explicit, named paramaters, and the map is just for the output
mime types. This is in service of the code cell transformer eventually
having a different base class (which is needed to support directives).
When all the transformers were in a single map, it was not possible to
express a different type for one key in the type system.