Implement basic code folding
This required transformers that can take directives. I introduced a new
Transformer base class, `TransformerWithDirectives` that does this and
made this the type for the Python source transfomer. I think implemented
the Cold Fold transformer based on this.

Because transformers are stateless and I didn't want to change the method
signatures to take directives, TransformerWithDirectives offers a method
that takes directives, sets them internally, and returns a context manager
that guarantees they'll be cleared when the context exits.