Allow ":" in value part of directives
Directives appear in comments in the form:

```
```

However the value part should be able to contain a colon. Previously, I
was splitting the string on the colon character and expecting only two
parts to come out of that. If there's a colon within value, you'd get
more parts. I've changed it to search for the first colon and treat
everything after it as part of the value.