Support `hide` directive for code cells
Per https://nbdev.fast.ai/explanations/directives.html#hide,
the hide directive should cause the source code and output
to be omitted when rendering. This is equivalent to

```
```

In implementing this, I refactored the interpretation of
directives into a standalone function and added tests. I
also added an example of hide in the test case notebook
and re-ran the nbs/markdown/03_markdown_api.ipynb cell that
exercises it.